# Audit the use of unsafe in uri/path.rs