@sbosnick 2020-04-17
If there are any concerns about a possible performance impact I would be happy to add benchmarks to answer that question.