@sbosnick 2020-05-15
The latest commit removes the performance regression that the benchmarks in `benches/uri.rs` showed (to within the margin of error). This retains the one-loop version of the core parsing code but changed from internal iteration (using `try_fold()`) to external iteration.

It appears that the outstanding issue then is whether this code should use a two-loop version that emphasizes parsing for two separate things (the path and the query) of a one-loop version that emphasized checking each  byte in turn that will eventually be passed to `ByteStr::from_utf8_unchecked()`.

I favour the latter because it makes the soundness easier to see. I can convert this back to a two-loop version, though, if there is a reason to prefer that approach.